workaround to avoid provider url collision with openai provider in main moly app
