+++ src/data/mod.rs
@@ -4,6 +4,7 @@ pub mod deep_inquire_client;
 pub mod downloads;
 pub mod moly_client;
 pub mod openai_client;
+pub mod openai_image_client;
 pub mod preferences;
 pub mod providers;
 pub mod search;
